Remove unnecessary squeeze operation for single input
Shouldn't have used a squeeze like this conditional on whether it is being used for full sequence or not. That should be handled elsewhere where required, otherwise this fails when using e.g. step size 1 (or in my case while im initting the network).
Probably breaks the included training code but that isn't what this fork is for so...